test/correctness/requirements.txt
pymongo==3.6.1
python-dateutil
PyYaml
psutil
pytest
coloredlogs==4.0.0
